% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CalculateFontSizeForBars}
\alias{CalculateFontSizeForBars}
\title{Calculates fontsize of value labels for barplots}
\usage{
CalculateFontSizeForBars(nbars)
}
\arguments{
\item{nbars}{Number of bars in barplot
CalculateFontSizeForBars()}
}
\description{
Calculates fontsize of value labels for barplots. Values found using trial and error.
}
