% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CalculateNs}
\alias{CalculateNs}
\title{Calculate the sample size of dataFrame for each subset}
\usage{
CalculateNs(dataFrame, columnName)
}
\arguments{
\item{dataFrame}{Contains filtered data, i.e. the data rows that will be used for plot.}

\item{columnName}{Name of column containing subsets.}
}
\value{
dataFrame containing new column called rowcount that has counts for each unique value in original dataFrame.
CalculateNs()
}
\description{
Calculate the sample size of dataFrame, i.e. How many times each unique value appears in columName. 
E.g. a column called gender would contain Male / Female text. Counts number of males and number of 
females. Returns result as a dataFrame.
}
