% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CalculatePctsForDemog}
\alias{CalculatePctsForDemog}
\title{Summarizes percentage of answers according to a demography}
\usage{
CalculatePctsForDemog(weightedData, dataFrame, demography)
}
\description{
Summarizes percentage of answers according to a demography. E.g. Demography column is called gender, 
containing male/female answers. Brand column contains different brands of beers. We get out a dataFrame 
with beer brands: how many males knew the beer brand and how many females knew the beer brand.
}
\examples{
CalculatePctsForDemog()
}
