% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CalculateTop3}
\alias{CalculateTop3}
\title{Calculates the sum of all positive answers of a question with a scale.}
\usage{
CalculateTop3(dataFrame)
}
\arguments{
\item{dataFrame}{Contains columns Scale, XLabels, pct}
}
\value{
tidyDataPos A dataFrame containing top3 summary data
CalculateTop3()
}
\description{
Scale value have to be in a column called "Scale". The x value labels have to be in a colmun called "XLabels". 
Answer percentages have to be in column called "pct". Verified that this function works with scale that has 
an odd number of choices. May not work with a scale that has an even number of choices.
}
