% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{ChangeValue}
\alias{ChangeValue}
\title{Changes values in a dataFrame}
\usage{
ChangeValue(dataFrame, searchfor, replacement)
}
\arguments{
\item{dataFrame}{The dataframe to be edited}

\item{searchfor}{The value we want to replace}

\item{replacement}{The value that will be used as replacement}
}
\value{
dataFrame in a form that is ready for plotting
ChangeValue()
}
\description{
Goes through all elements in a dataFrame and replaces what is being searched for with a replacement.
}
