% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CheckZeroesAndOnes}
\alias{CheckZeroesAndOnes}
\title{Checks if first column and row contains zeroes and ones}
\usage{
CheckZeroesAndOnes(dataFrame)
}
\description{
Checks if first column and first row of dataFrame contains zeroes and ones. If yes, it can be assumed that all of the data is zeroes and ones.
Excludes column "weight".
CheckZeroesAndOnes
}
