% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{CleanDataFrame}
\alias{CleanDataFrame}
\title{Checks if dataFrame contains column weight and adds rowcount column}
\usage{
CleanDataFrame(dataFrame)
}
\description{
Checks if dataFrame contains column weight. If not, it will add weight column containing 1 for each row. This function will also add a rowcount column containing ones to be used for nrow calculations.
}
\examples{
CleanDataFrame()
}
