% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muensing.R
\name{ConvertToZeroesAndOnes}
\alias{ConvertToZeroesAndOnes}
\title{Converts a dataFrame containing text into Zeroes and Ones
Converts a dataFrame containing text into Zeroes and Ones. Dataframe should contain one column with text and another column with weights.}
\usage{
ConvertToZeroesAndOnes(dataFrame)
}
\arguments{
\item{dataFrame}{Two colmuns: first containing text, second containing values/weights}
}
\value{
zeroesAndOnesdf DataFrame in zeroes and ones form.
}
\description{
Converts a dataFrame containing text into Zeroes and Ones
Converts a dataFrame containing text into Zeroes and Ones. Dataframe should contain one column with text and another column with weights.
}
\examples{
ConvertToZeroesAndOnes()
}
